refactored code, but with broken tests
