+++ mapping_suite_sdk/mapping_package_v3/models/mapping_package_v3_lightweight.py
